requirements.txt
Django
djangorestframework
reportlab
python-dotenv
django-debug-toolbar
django-mptt